Allow the user to define 'quiet hours' for notifications.
Introduces a notion of user-defined 'quiet hours'. During this period,
notifications will be: (1) silenced, (2) have vibrations removed, and
(3) have their requested colors muted (to dim the LED on capable hardware).

Note that (3) works well on Nexus One and gracefully degrades on sapphire.
That is, LED colors are reduced, but hardware still displays them the same.

Bug: 2164
Bug: 2252
Change-Id: I66b4dbab8ccc30741b15629fb713ff18b4511b47